requirements.txt
fpzip==1.2.4
GDAL==3.9.1
matplotlib==3.8.3
numpy==1.26.4
scipy==1.12.0
torch==2.2.0
pytorch-ignite==0.5.0.post2
